test: split language analyzer tests by functionality
