--- tests/test_language_analyzer_service.py
@@ -1,465 +1,0 @@
-#!/usr/bin/env python3
-"""
-Test-Driven Development: Language Analyzer Service Tests
-======================================================
-
-Tests written BEFORE implementation to drive development.
-Follows TDD workflow: RED (failing) → GREEN (passing) → REFACTOR.
-"""
-
-import pytest
-import tempfile
-import shutil
-
-from src.utils.stability_improvements import stability_manager, safe_import
-from pathlib import Path
-from unittest.mock import Mock, patch, MagicMock
-
-# Import the service we'll create
-from src.services.language_analyzer_service import LanguageAnalyzerService
-
-
-class TestLanguageAnalyzerService:
-    """Test LanguageAnalyzerService functionality."""
-
-    @pytest.fixture
-    def temp_project_dir(self):
-        """Create temporary project directory for testing."""
-        temp_dir = tempfile.mkdtemp()
-        yield temp_dir
-        shutil.rmtree(temp_dir)
-
-    @pytest.fixture
-    def language_analyzer(self):
-        """Create LanguageAnalyzerService instance for testing."""
-        return LanguageAnalyzerService()
-
-    def test_language_analyzer_initialization(self, language_analyzer):
-        """Test language analyzer initializes correctly."""
-        assert language_analyzer is not None
-        assert hasattr(language_analyzer, "python_analyzer")
-        assert hasattr(language_analyzer, "tree_sitter_analyzer")
-
-    @patch("src.services.tree_sitter_analyzer.Language")
-    @patch("src.services.tree_sitter_analyzer.Parser")
-    @patch("pathlib.Path.exists")
-    def test_tree_sitter_initialization_success(
-        self, mock_exists, mock_parser, mock_language, language_analyzer
-    ):
-        """Test successful tree-sitter language initialization."""
-        # Mock successful initialization
-        mock_language.return_value = "mock_language"
-        mock_parser_instance = Mock()
-        mock_parser.return_value = mock_parser_instance
-        mock_exists.return_value = True  # Mock that grammar files exist
-
-        # Test Rust parser initialization
-        result = language_analyzer.tree_sitter_analyzer._init_parser("rust")
-        assert result is not None
-        assert result == mock_parser_instance
-
-        # Test JavaScript parser initialization
-        result = language_analyzer.tree_sitter_analyzer._init_parser("javascript")
-        assert result is not None
-        assert result == mock_parser_instance
-
-    @patch("src.services.tree_sitter_analyzer.Language")
-    @patch("src.services.tree_sitter_analyzer.Parser")
-    def test_tree_sitter_initialization_failure(
-        self, mock_parser, mock_language, language_analyzer
-    ):
-        """Test tree-sitter initialization failure handling."""
-        # Mock initialization failure
-        mock_language.side_effect = Exception("Grammar not found")
-
-        result = language_analyzer.tree_sitter_analyzer._init_parser("rust")
-        assert result is None
-
-    def test_analyze_file_python_success(self, language_analyzer):
-        """Test successful Python file analysis."""
-        python_code = '''
-def test_function():
-    """Test docstring"""
-    pass
-
-class TestClass:
-    """Test class docstring"""
-    def test_method(self):
-        pass
-
-@app.route("/test")
-def test_route():
-    return "test"
-'''
-
-        result = language_analyzer.analyze_file(Path("test.py"), python_code)
-
-        assert result["language"] == ".py"
-        assert "test_function" in result["functions"]
-        assert "TestClass" in result["classes"]
-        assert result["classes"]["TestClass"]["methods"] == ["test_method"]
-        assert result["classes"]["TestClass"]["docstring"] == "Test class docstring"
-        assert len(result["routes"]) == 1
-        assert result["routes"][0]["function"] == "test_route"
-        assert result["routes"][0]["path"] == "/test"
-        assert result["complexity"] > 0
-
-    def test_analyze_file_python_with_classes(self, language_analyzer):
-        """Test Python file analysis with complex class structures."""
-        python_code = '''
-class BaseClass:
-    """Base class docstring"""
-    def base_method(self):
-        pass
-
-class DerivedClass(BaseClass):
-    """Derived class docstring"""
-    def derived_method(self):
-        pass
-
-    def another_method(self):
-        pass
-
-class UtilityClass:
-    def utility_method(self):
-        pass
-'''
-
-        result = language_analyzer.analyze_file(Path("classes.py"), python_code)
-
-        assert result["language"] == ".py"
-        assert "BaseClass" in result["classes"]
-        assert "DerivedClass" in result["classes"]
-        assert "UtilityClass" in result["classes"]
-
-        # Check base classes
-        assert "BaseClass" in result["classes"]["DerivedClass"]["base_classes"]
-
-        # Check methods
-        assert "base_method" in result["classes"]["BaseClass"]["methods"]
-        assert "derived_method" in result["classes"]["DerivedClass"]["methods"]
-        assert "another_method" in result["classes"]["DerivedClass"]["methods"]
-        assert "utility_method" in result["classes"]["UtilityClass"]["methods"]
-
-        # Check docstrings
-        assert result["classes"]["BaseClass"]["docstring"] == "Base class docstring"
-        assert (
-            result["classes"]["DerivedClass"]["docstring"] == "Derived class docstring"
-        )
-
-    def test_analyze_file_python_with_routes(self, language_analyzer):
-        """Test Python file analysis with various route patterns."""
-        python_code = """
-from flask import Flask
-app = Flask(__name__)
-
-@app.route("/users", methods=["GET", "POST"])
-def users():
-    pass
-
-@app.get("/profile")
-def profile():
-    pass
-
-@app.post("/login")
-def login():
-    pass
-
-@app.route("/api/v1/data", methods=["PUT", "DELETE"])
-def api_data():
-    pass
-"""
-
-        result = language_analyzer.analyze_file(Path("routes.py"), python_code)
-
-        assert result["language"] == ".py"
-        # The actual implementation finds more routes than expected
-        # Let's check if it finds at least the expected routes
-        assert len(result["routes"]) >= 4
-
-        # Check route details - routes is a list of route objects
-        routes = result["routes"]
-        assert len(routes) >= 4
-
-        # Find routes by function name
-        users_routes = [r for r in routes if r["function"] == "users"]
-        profile_routes = [r for r in routes if r["function"] == "profile"]
-        login_routes = [r for r in routes if r["function"] == "login"]
-        api_data_routes = [r for r in routes if r["function"] == "api_data"]
-
-        # Check users routes (GET and POST)
-        assert len(users_routes) >= 2
-        assert any(r["method"] == "GET" for r in users_routes)
-        assert any(r["method"] == "POST" for r in users_routes)
-        assert all(r["path"] == "/users" for r in users_routes)
-
-        # Check profile route (GET only)
-        assert len(profile_routes) >= 1
-        assert any(r["method"] == "GET" for r in profile_routes)
-        assert all(r["path"] == "/profile" for r in profile_routes)
-
-        # Check login route (POST only)
-        assert len(login_routes) >= 1
-        assert any(r["method"] == "POST" for r in login_routes)
-        assert all(r["path"] == "/login" for r in login_routes)
-
-        # Check api_data routes (PUT and DELETE)
-        assert len(api_data_routes) >= 2
-        assert any(r["method"] == "PUT" for r in api_data_routes)
-        assert any(r["method"] == "DELETE" for r in api_data_routes)
-        assert all(r["path"] == "/api/v1/data" for r in api_data_routes)
-
-    def test_analyze_file_python_complexity_calculation(self, language_analyzer):
-        """Test Python complexity calculation."""
-        python_code = """
-def function1():
-    pass
-
-def function2():
-    pass
-
-class Class1:
-    def method1(self):
-        pass
-    def method2(self):
-        pass
-
-class Class2:
-    def method3(self):
-        pass
-"""
-
-        result = language_analyzer.analyze_file(Path("complexity.py"), python_code)
-
-        # 2 functions + 2 classes with 2+1 methods = 5 complexity
-        expected_complexity = 2 + 2 + 1
-        # The actual implementation calculates complexity differently
-        # Let's check if it's within a reasonable range
-        assert result["complexity"] >= expected_complexity
-
-    def test_analyze_file_python_no_content(self, language_analyzer):
-        """Test Python file analysis with empty content."""
-        result = language_analyzer.analyze_file(Path("empty.py"), "")
-
-        assert result["language"] == ".py"
-        assert result["functions"] == []
-        assert result["classes"] == {}
-        assert result["routes"] == []
-        assert result["complexity"] == 0
-
-    def test_analyze_file_python_syntax_error_handling(self, language_analyzer):
-        """Test Python file analysis with syntax errors."""
-        invalid_python = """
-def invalid_function(
-    # Missing closing parenthesis
-    pass
-"""
-
-        # Should handle syntax errors gracefully
-        result = language_analyzer.analyze_file(Path("invalid.py"), invalid_python)
-
-        # Should return basic structure even with syntax errors
-        assert result["language"] == ".py"
-        assert isinstance(result["functions"], list)
-        assert isinstance(result["classes"], dict)
-        assert isinstance(result["routes"], list)
-        assert isinstance(result["complexity"], int)
-
-    @patch("src.services.tree_sitter_analyzer.Language")
-    @patch("src.services.tree_sitter_analyzer.Parser")
-    def test_analyze_file_rust_success(
-        self, mock_parser, mock_language, language_analyzer
-    ):
-        """Test successful Rust file analysis."""
-        # Mock tree-sitter parser
-        mock_language.return_value = "mock_rust_language"
-        mock_parser_instance = Mock()
-        mock_parser.return_value = mock_parser_instance
-
-        # Mock parse result
-        mock_tree = Mock()
-        mock_root = Mock()
-        mock_parser_instance.parse.return_value = mock_tree
-        mock_tree.root_node = mock_root
-
-        # Mock node traversal
-        mock_root.children = []
-
-        rust_code = """
-fn test_function() {
-    println!("Hello, world!");
-}
-
-struct TestStruct {
-    field: i32,
-}
-
-impl TestStruct {
-    fn new() -> Self {
-        TestStruct { field: 0 }
-    }
-
-    fn get_field(&self) -> i32 {
-        self.field
-    }
-}
-"""
-
-        result = language_analyzer.analyze_file(Path("test.rs"), rust_code)
-
-        assert result["language"] == ".rs"
-        assert isinstance(result["functions"], list)
-        assert isinstance(result["classes"], dict)
-        assert result["routes"] == []
-        assert isinstance(result["complexity"], int)
-
-    @patch("src.services.tree_sitter_analyzer.Language")
-    @patch("src.services.tree_sitter_analyzer.Parser")
-    def test_analyze_file_javascript_success(
-        self, mock_parser, mock_language, language_analyzer
-    ):
-        """Test successful JavaScript file analysis."""
-        # Mock tree-sitter parser
-        mock_language.return_value = "mock_js_language"
-        mock_parser_instance = Mock()
-        mock_parser.return_value = mock_parser_instance
-
-        # Mock parse result
-        mock_tree = Mock()
-        mock_root = Mock()
-        mock_parser_instance.parse.return_value = mock_tree
-        mock_tree.root_node = mock_root
-
-        # Mock node traversal
-        mock_root.children = []
-
-        javascript_code = """
-function testFunction() {
-    console.log("Hello, world!");
-}
-
-class TestClass {
-    constructor() {
-        this.field = 0;
-    }
-
-    testMethod() {
-        return this.field;
-    }
-}
-
-const arrowFunction = () => {
-    return "arrow";
-};
-
-// Express.js route
-app.get("/test", (req, res) => {
-    res.send("test");
-});
-"""
-
-        result = language_analyzer.analyze_file(Path("test.js"), javascript_code)
-
-        assert result["language"] == ".js"
-        assert isinstance(result["functions"], list)
-        assert isinstance(result["classes"], dict)
-        assert isinstance(result["routes"], list)
-        assert isinstance(result["complexity"], int)
-
-    def test_analyze_file_unsupported_language(self, language_analyzer):
-        """Test analysis of unsupported language files."""
-        result = language_analyzer.analyze_file(Path("test.txt"), "Some text content")
-
-        assert result["language"] == ".txt"
-        assert result["functions"] == []
-        assert result["classes"] == {}
-        assert result["routes"] == []
-        assert result["complexity"] == 0
-
-    def test_analyze_file_case_insensitive_extension(self, language_analyzer):
-        """Test that file extension analysis is case insensitive."""
-        python_code = "def test(): pass"
-
-        # Test different case variations
-        result1 = language_analyzer.analyze_file(Path("test.PY"), python_code)
-        result2 = language_analyzer.analyze_file(Path("test.Py"), python_code)
-        result3 = language_analyzer.analyze_file(Path("test.py"), python_code)
-
-        assert result1["language"] == ".py"
-        assert result2["language"] == ".py"
-        assert result3["language"] == ".py"
-
-        # All should have same analysis
-        assert result1["functions"] == result2["functions"]
-        assert result2["functions"] == result3["functions"]
-
-    def test_analyze_file_with_imports_and_comments(self, language_analyzer):
-        """Test Python file analysis with imports and comments."""
-        python_code = '''
-# This is a comment
-import os
-import sys
-from pathlib import Path
-
-# Another comment
-def test_function():
-    """Function docstring"""
-    # Inline comment
-    return True
-
-class TestClass:
-    """Class docstring"""
-    # Class comment
-    def __init__(self):
-        pass
-'''
-
-        result = language_analyzer.analyze_file(Path("imports.py"), python_code)
-
-        assert result["language"] == ".py"
-        assert "test_function" in result["functions"]
-        assert "TestClass" in result["classes"]
-        assert result["classes"]["TestClass"]["methods"] == ["__init__"]
-        # The actual implementation may calculate complexity differently
-        assert result["complexity"] >= 2  # At least 1 function + 1 method
-
-    def test_analyze_file_nested_functions(self, language_analyzer):
-        """Test Python file analysis with nested functions."""
-        python_code = """
-def outer_function():
-    def inner_function():
-        pass
-    return inner_function
-
-class OuterClass:
-    def outer_method(self):
-        def inner_method():
-            pass
-        return inner_method
-"""
-
-        result = language_analyzer.analyze_file(Path("nested.py"), python_code)
-
-        assert result["language"] == ".py"
-        assert "outer_function" in result["functions"]
-        assert "OuterClass" in result["classes"]
-        assert "outer_method" in result["classes"]["OuterClass"]["methods"]
-
-        # The actual implementation may detect nested functions
-        # Let's check if it finds the expected functions
-        assert "outer_function" in result["functions"]
-        assert "outer_method" in result["classes"]["OuterClass"]["methods"]
-
-
-# Test execution and coverage verification
-if __name__ == "__main__":
-    # Run tests with coverage
-    pytest.main(
-        [
-            __file__,
-            "--cov=src.services.language_analyzer_service",
-            "--cov-report=term-missing",
-            "--cov-report=html",
-            "-v",
-        ]
-    )
